rename spot -> ledspot
